64-bit changes: casts, and ioctl cmds are u_longs
SCCS-vsn: sys/net/bpf.c 8.4
SCCS-vsn: sys/net/if.c 8.5
SCCS-vsn: sys/net/bpf.h 8.2
SCCS-vsn: sys/net/if.h 8.2
SCCS-vsn: sys/net/route.c 8.3
SCCS-vsn: sys/net/if_sl.c 8.9
SCCS-vsn: sys/net/route.h 8.4
SCCS-vsn: sys/net/if_loop.c 8.2
SCCS-vsn: sys/net/if_slvar.h 8.4

+++ usr/src/sys/net/bpf.c
@@ -9,7 +9,7 @@
  *
  * %sccs.include.redist.c%
  *
- *      @(#)bpf.c	8.3 (Berkeley) %G%
+ *      @(#)bpf.c	8.4 (Berkeley) %G%
  *
  * static char rcsid[] =
  * "$Header: bpf.c,v 1.33 91/10/27 21:21:58 mccanne Exp $";
@@ -558,7 +558,7 @@ reset_d(d)
 int
 bpfioctl(dev, cmd, addr, flag)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t addr;
 	int flag;
 {
